chore: add footer in index page
