+++ cosmonaut-frontend/src/layout/Index/IndexInitialPage.jsx
@@ -5,6 +5,7 @@ import Navbar from "../../components/Navbar/Navbar";
 import Video from "../../assets/bg2-v5.mp4";
 import { indexInfo } from "../../core/config/indexInfo";
 import { ReadyJourney } from "./components/ReadyJourney";
+import Footer from "../../components/Footer/Footer";
 
 const Title = tw.h2`text-2xl md:text-4xl text-center lg:text-left mt-2 text-orange-400 lg:mb-8 mb-6 font-heading`;
 const LessonList = tw.div`md:space-y-5 space-y-3 md:mx-0 mx-6`;
@@ -23,7 +24,7 @@ function IndexInitialPage() {
   return (
     <>
       <Navbar />
-      <div className="h-[860px] overflow-y-scroll z-0 relative lg:pb-20 bg-cover bg-center bg-opacity-10 lg:pt-32">
+      <div className="h-[880px] overflow-y-scroll z-0 relative lg:pb-20 bg-cover bg-center bg-opacity-10 lg:pt-32">
         <div className="z-[-1] absolute bottom-0 w-full h-auto">
           <video autoPlay muted loop playsInline>
             <source src={Video} type="video/mp4" />
@@ -71,6 +72,7 @@ function IndexInitialPage() {
             </Curriculum>
           </div>
         </div>
+        <Footer />
       </div>
     </>
   );
