+++ cosmonaut-frontend/src/layout/Index/IndexPage.jsx
@@ -10,6 +10,7 @@ import { useEffect } from "react";
 import Video from "../../assets/bg2-v5.mp4";
 import { handleModalAtom } from "../../core/state/handleModalState";
 import { indexInfo } from "../../core/config/indexInfo";
+import Footer from "../../components/Footer/Footer";
 
 const Curriculum = tw.div`w-full mb-14 lg:mb-0 lg:col-span-1 col-span-2 lg:order-2 order-1`;
 const Title = tw.h2`text-2xl md:text-4xl text-center lg:text-left mt-2 text-orange-400 lg:mb-8 mb-6 font-heading`;
@@ -110,7 +111,7 @@ function IndexPage() {
   return (
     <>
       <Navbar />
-      <div className="h-[860px] overflow-y-scroll z-0 relative lg:pb-20 bg-cover bg-center bg-opacity-10 lg:pt-32">
+      <div className="h-[880px] overflow-y-scroll z-0 relative lg:pb-20 bg-cover bg-center bg-opacity-10 lg:pt-32">
         <div className="z-[-1] absolute bottom-0 w-full h-auto">
           <video autoPlay muted loop playsInline>
             <source src={Video} type="video/mp4" />
@@ -178,6 +179,7 @@ function IndexPage() {
             </Curriculum>
           </div>
         </div>
+        <Footer />
       </div>
     </>
   );
